Terminado y subido a Firebase

+++ src/app/edu-exp/edu-exp.component.ts
@@ -55,11 +55,25 @@ export class EduExpComponent implements OnInit {
   // Borrar experiencia
 
   deleteExp(id: number) {
-    this.expService.deleteExperience(id).subscribe(
-      data => {
-        console.log(data);
-      });
-      this.listExperiences();
+    Alerta.fire({
+      title: '¿Seguro que querés eliminar este elemento?',
+      text: 'Los elementos eliminados no pueden recuperarse',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'No, no quiero eliminar',
+    }).then((result) => {
+      if(result.value) {
+        this.expService.deleteExperience(id).subscribe(
+          data => {
+          Alerta.fire('Removido', 'Elemento removido con éxito. Recarga la página', 'success')
+          console.log(data);
+          });
+          this.listExperiences();
+      } else if (result.dismiss === Alerta.DismissReason.cancel) {
+        Alerta.fire('Cancelado', 'Operación cancelada', 'error')
+      }
+    })
   }
 
   // Educación
